partiicpant view per week
* Added week selector on participant view to filter by week.
* Added days column that shows only the days participant will attend for sleected week.
* Added paid in full and amount paid columns.
* Added app bar for the admin pages.
* Added ability to export data grids as csv with custom file names.
* Fixes issue #16
* Fixes issue #15
* Fixes issue #13
* Fixes issue #14
