--- SummerCamp/app/participants/page.tsx
@@ -1,98 +1,0 @@
-"use client";
-import { useState, useEffect } from "react";
-import { format } from 'date-fns';
-import { DataGrid, GridColDef, GridFilterOperator } from '@mui/x-data-grid';
-import Paper from '@mui/material/Paper';
-import Button from '@mui/material/Button';
-import Link from 'next/link';
-import '../globals.css';
-
-export default function App() {
-
-    const columns: GridColDef[] = [
-        { field: 'first_name', headerName: 'First Name', flex: 2 },
-        { field: 'last_name', headerName: 'Last name', flex: 2 },
-        { field: 'age', headerName: 'Age', flex: 1, type: 'number' },
-        {
-            field: 'gender',
-            headerName: 'Gender',
-            type: 'string',
-            flex: 1,
-        },
-        {
-            field: 'dates',
-            headerName: 'Dates',
-            description: 'This column has a value getter and is not sortable.',
-            sortable: false,
-            flex: 3,
-            type: 'date',
-            valueGetter: (value, row) => row.dates.map((date: string) => new Date(date)),
-            valueFormatter: (value, row) => {
-                const dates: Date[] = row.dates;
-                return dates.map(d => format(d, 'MM/dd/yyyy')).join(', ');
-            },
-        },
-        {
-            field: 'details',
-            headerName: 'Details',
-            sortable: false,
-            flex: 1,
-            minWidth: 100,
-            renderCell: (params: any) => {
-                const id = params?.row?.id;
-                return (
-                    <Link href={`/participants/${id}`}>
-                        <Button variant="contained" size="small">View Details</Button>
-                    </Link>
-                );
-            },
-        },
-    ];
-
-    const paginationModel = {  };
-
-    const [participants, setParticipants] = useState<any[]>([]);
-
-    useEffect(() => {
-        let mounted = true;
-        const controller = new AbortController();
-
-        async function load() {
-            try {
-                const response = await fetch("http://localhost:8080/participants", { signal: controller.signal });
-                if (!response.ok) {
-                    console.error("Failed to fetch participants", response.status);
-                    return;
-                }
-                const data = await response.json();
-                if (mounted) setParticipants(data);
-            } catch (err: any) {
-                if (err.name === 'AbortError') return;
-                console.error("Error fetching participants:", err);
-            }
-        }
-
-        load();
-
-        return () => {
-            mounted = false;
-            controller.abort();
-        };
-    }, []);
-
-    return (
-        <div className="content admin">
-            <h1>Participants</h1>
-            <Paper sx={{ height: '100%', width: '80%' }}>
-                <DataGrid
-                    rows={participants}
-                    columns={columns}
-                    initialState={{ pagination: { paginationModel } }}
-                    pageSizeOptions={[5, 10]}
-                    checkboxSelection
-                    sx={{ border: 0 }}
-                />
-            </Paper>
-        </div>
-    );
-}
