+++ SummerCamp/package.json
@@ -9,7 +9,10 @@
     "lint": "eslint"
   },
   "dependencies": {
+    "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.7",
+    "@mui/material": "^7.3.7",
     "@mui/x-data-grid": "^8.25.0",
     "date-fns": "^4.1.0",
     "next": "^16.1.1",
